requirements.txt
cachetools==4.2.4
diskcache==5.4.0
dpath==2.0.6
einops==0.6.1
h5py==3.6.0
icecream==2.1.2
ipython==8.0.1
lpips==0.1.4
matplotlib==3.5.1
numpy==1.22.2
opencv_python==4.5.5.62
opencv_transforms==0.0.6
pandas==1.1.5
Pillow==10.0.1
plotly==5.6.0
positional_encodings==6.0.1
pytesseract==0.3.8
python_Levenshtein==0.21.1
pytorch_lightning==1.6.4
pytorch_msssim==0.2.1
PyYAML==6.0
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.8.0
timm==0.5.4
tqdm==4.62.3
unflatten==0.1.1
vit_pytorch==0.38.1
gdown==4.7.1
types-PyYAML==6.0.12.12
ipykernel==6.26.0
kaleido==0.2.1
mypy==1.7.1
pre-commit==3.5.0
ipywidgets==8.1.1